+++ src/lib/pages/LinksTab.svelte
@@ -5,15 +5,17 @@
 
 <!-- Links ImageSrc={[file-name]ScreenshotUrl} -->
     
-<h1>Socials and Websites</h1>
+<h1>Links</h1>
 <div class="container">
     <ul>
-        <h2>Socials</h2>
+        <h2>Socials/GitHub</h2>
         <li><a a href="https://discord.gg/jvmadKWFe9" target='_blank' class='text-blue-600 underline'>Official English Discord Server</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
         <li><a a href="https://discord.gg/CJ4nTkpy7t" target='_blank' class='text-blue-600 underline'>Official Japanese Discord Server</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
         <li><a a href="https://www.youtube.com/@OpenTaiko" target='_blank' class='text-blue-600 underline'>Official YouTube Channel</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
         <li><a a href="https://bsky.app/profile/opentaiko.bsky.social" target='_blank' class='text-blue-600 underline'>Official Bluesky Account</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
         <li><a a href="https://x.com/OpenTaiko" target='_blank' class='text-blue-600 underline'>Official Twitter/X Account</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://github.com/0auBSQ/OpenTaiko" target='_blank' class='text-blue-600 underline'>Main OpenTaiko Source Code</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://github.com/OpenTaiko/OpenTaiko-Hub" target='_blank' class='text-blue-600 underline'>OpenTaiko Hub Source Code</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
         <li><a a href="https://github.com/OpenTaiko" target='_blank' class='text-blue-600 underline'>GitHub Organization</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
     </ul>
     <img src={optkkunUrl} alt="OpenTaiko-kun" class="image optkkun"/>