+++ src/lib/pages/ToolsTab.svelte
@@ -12,7 +12,7 @@
     import tjatoolsScreenshotUrl from '$lib/tjatools.png';
 	import subeditScreenshotUrl from '$lib/subedit.png';
 
-    let currentTool = 0;
+    let currentTab = 0;
 	let optk_OS = "Win";
 
 	onMount(async () => {
@@ -30,18 +30,22 @@
 	border=""
 	class="bg-surface-100-800-token w-full"
 	>
-	<Tab bind:group={currentTool} name="tab1" value={0}>
+	<Tab bind:group={currentTab} name="tab1" value={0}>
 		<svelte:fragment slot="lead"><i class="fa-regular fa-file-audio"></i></svelte:fragment>
 		<span>Charting</span>
 	</Tab>
-	<Tab bind:group={currentTool} name="tab2" value={1}>
+	<Tab bind:group={currentTab} name="tab2" value={1}>
+		<svelte:fragment slot="lead"><i class="fa-solid fa-music"></i></svelte:fragment>
+		<span>Lyrics</span>
+	</Tab>
+	<Tab bind:group={currentTab} name="tab3" value={2}>
 		<svelte:fragment slot="lead"><i class="fa-solid fa-envelope-open-text"></i></svelte:fragment>
 		<span>Submit your content!</span>
 	</Tab>
 	<!-- ... -->
 </TabGroup>
 <!-- Charting -->
-{#if currentTool === 0}
+{#if currentTab === 0}
 	<h2>Charting tutorial by ugyuu:</h2>
 	<lite-youtube width="100%" videoid="U0i-z-tpxY8" playlabel="Play: Keynote (Google I/O '18)"></lite-youtube>
 	{#if optk_OS === "Win"}
@@ -80,6 +84,11 @@
 			CardSubtitle="Official Latest"
 			CardText ="The recommended tool to sync your charts trouble free!"
 			/>
+		{/if}
+	</div>
+{/if}
+{#if currentTab === 1}
+	<div class="w-full text-token grid grid-cols-1 md:grid-cols-2 gap-4">
 		<ToolCard 
 			Url="https://github.com/SubtitleEdit/subtitleedit"
 			ImageSrc={subeditScreenshotUrl}
@@ -87,11 +96,10 @@
 			CardSubtitle="Official Latest"
 			CardText ="The recommended tool to make VTT files for your charts!"
 			/>
-		{/if}
 	</div>
 {/if}
 <!-- Submit your content -->
-{#if currentTool === 1}
+{#if currentTab === 2}
 	<div class="w-full text-token grid grid-cols-1 md:grid-cols-2 gap-4">
 		<ToolCard 
 			Url="https://forms.gle/WXNUwjJyLdJoeRSM6"